chore: rename example
