+++ apps/src/tests/index.ts
@@ -142,4 +142,4 @@ export { default as TestMemoryLeak } from './TestMemoryLeak';
 export { default as TestFormSheet } from './TestFormSheet';
 export { default as TestAndroidTransitions } from './TestAndroidTransitions';
 export { default as TestAnimation } from './TestAnimation';
-export { default as TestSheetTranslate } from './TestSheetTranslate';
+export { default as TestSheetFeatures } from './TestSheetFeatures';
